Select referee sheet dynamically, move shared referee components
