+++ front-end/src/apps/Referee/BlueReferee.tsx
@@ -3,15 +3,26 @@ import MatchStateListener from 'src/components/MatchStateListener/MatchStateList
 import MatchUpdateListener from 'src/components/MatchUpdateListener/MatchUpdateListener';
 import PrestartListener from 'src/components/PrestartListener/PrestartListener';
 import DefaultLayout from 'src/layouts/DefaultLayout';
-import BlueScoreSheet from './components/games/HyodrogenHorizons/BlueScoreSheet';
+import { useSeasonComponents } from '@seasons/index';
+import { useRecoilValue } from 'recoil';
+import { currentEventSelector } from '@stores/NewRecoil';
 
 const BlueReferee: FC = () => {
+  const event = useRecoilValue(currentEventSelector);
+  if (!event) {
+    return 'No event selected';
+  }
+  const seasonComponents = useSeasonComponents(event.seasonKey);
+  if (!seasonComponents) {
+    return 'Unknown season';
+  }
+
   return (
     <DefaultLayout containerWidth='xl'>
       <PrestartListener />
       <MatchStateListener />
       <MatchUpdateListener />
-      <BlueScoreSheet />
+      <seasonComponents.RefereeScoreSheet alliance='blue' />
     </DefaultLayout>
   );
 };
