+++ front-end/src/apps/Referee/HeadReferee.tsx
@@ -3,8 +3,8 @@ import { useRecoilValue } from 'recoil';
 import DefaultLayout from 'src/layouts/DefaultLayout';
 import { matchInProgressAtom } from '@stores/NewRecoil';
 import { Box } from '@mui/material';
-import RefereeSheet from './components/games/CarbonCapture/RefereeSheet';
-import ScoreSheetSmall from './components/games/CarbonCapture/ScoreSheetSmall';
+import RefereeSheet from '@seasons/CarbonCapture/referee/RefereeSheet';
+import ScoreSheetSmall from '@seasons/CarbonCapture/referee/ScoreSheetSmall';
 import PrestartListener from 'src/components/PrestartListener/PrestartListener';
 import MatchStateListener from 'src/components/MatchStateListener/MatchStateListener';
 import MatchUpdateListener from 'src/components/MatchUpdateListener/MatchUpdateListener';
