+++ front-end/src/apps/Referee/ScoreKeeper.tsx
@@ -3,7 +3,7 @@ import MatchStateListener from 'src/components/MatchStateListener/MatchStateList
 import MatchUpdateListener from 'src/components/MatchUpdateListener/MatchUpdateListener';
 import PrestartListener from 'src/components/PrestartListener/PrestartListener';
 import DefaultLayout from 'src/layouts/DefaultLayout';
-import ScoreSheet from './components/games/CarbonCapture/Scoresheet';
+import ScoreSheet from '@seasons/CarbonCapture/referee/Scoresheet';
 
 const ScoreKeeper: FC = () => {
   return (
