--- front-end/src/apps/Referee/components/games/HyodrogenHorizons/BlueScoreSheet.tsx
@@ -1,8 +1,0 @@
-import { FC } from 'react';
-import ScoreSheet from './ScoreSheet';
-
-const BlueScoreSheet: FC = () => {
-  return <ScoreSheet alliance='blue' />;
-};
-
-export default BlueScoreSheet;
