--- front-end/src/apps/Referee/components/games/HyodrogenHorizons/RedScoreSheet.tsx
@@ -1,8 +1,0 @@
-import { FC } from 'react';
-import ScoreSheet from './ScoreSheet';
-
-const RedScoreSheet: FC = () => {
-  return <ScoreSheet alliance='red' />;
-};
-
-export default RedScoreSheet;
