+++ front-end/src/seasons/CarbonCapture/referee/CarbonLevelInput.tsx
