+++ front-end/src/seasons/ChargedUp/referee/AutoScoreSheet.tsx
@@ -7,10 +7,10 @@ import {
   Match,
   MatchParticipant
 } from '@toa-lib/models';
-import StateToggle from '../../StateToggle';
+import StateToggle from '@components/Referee/StateToggle';
 import { SetterOrUpdater, useRecoilState } from 'recoil';
-import { matchInProgressAtom } from 'src/stores/NewRecoil';
-import NumberInput from '../../NumberInput';
+import { matchInProgressAtom } from '@stores/NewRecoil';
+import NumberInput from '@components/Referee/NumberInput';
 
 interface Props {
   alliance: Alliance;
@@ -18,7 +18,7 @@ interface Props {
   onUpdate?: (match: Match<ChargedUpDetails>) => void;
 }
 
-const TeleScoreSheet: FC<Props> = ({ alliance, participants, onUpdate }) => {
+const AutoScoreSheet: FC<Props> = ({ alliance, participants, onUpdate }) => {
   const [match, setMatch]: [
     Match<ChargedUpDetails> | null,
     SetterOrUpdater<Match<ChargedUpDetails> | null>
@@ -37,10 +37,52 @@ const TeleScoreSheet: FC<Props> = ({ alliance, participants, onUpdate }) => {
     return newMatch;
   };
 
+  const getMobility = (station: number): number | undefined => {
+    switch (station) {
+      case 11:
+        return match.details?.redAutoMobilityOne;
+      case 12:
+        return match.details?.redAutoMobilityTwo;
+      case 13:
+        return match.details?.redAutoMobilityThree;
+      case 21:
+        return match.details?.blueAutoMobilityOne;
+      case 22:
+        return match.details?.blueAutoMobilityTwo;
+      case 23:
+        return match.details?.blueAutoMobilityThree;
+      default:
+        return 0;
+    }
+  };
+
+  const updateMobility = (station: number, value: number) => {
+    switch (station) {
+      case 11:
+        setMatch(setDetails('redAutoMobilityOne', value));
+        break;
+      case 12:
+        setMatch(setDetails('redAutoMobilityTwo', value));
+        break;
+      case 13:
+        setMatch(setDetails('redAutoMobilityThree', value));
+        break;
+      case 21:
+        setMatch(setDetails('blueAutoMobilityOne', value));
+        break;
+      case 22:
+        setMatch(setDetails('blueAutoMobilityTwo', value));
+        break;
+      case 23:
+        setMatch(setDetails('blueAutoMobilityThree', value));
+        break;
+    }
+  };
+
   const handleTopPieceChange = (newValue: number) => {
     setMatch(
       setDetails(
-        alliance === 'red' ? 'redTeleTopPieces' : 'blueTeleTopPieces',
+        alliance === 'red' ? 'redAutoTopPieces' : 'blueAutoTopPieces',
         newValue
       )
     );
@@ -49,7 +91,7 @@ const TeleScoreSheet: FC<Props> = ({ alliance, participants, onUpdate }) => {
   const handleMidPieceChange = (newValue: number) => {
     setMatch(
       setDetails(
-        alliance === 'red' ? 'redTeleMidPieces' : 'blueTeleMidPieces',
+        alliance === 'red' ? 'redAutoMidPieces' : 'blueAutoMidPieces',
         newValue
       )
     );
@@ -58,7 +100,7 @@ const TeleScoreSheet: FC<Props> = ({ alliance, participants, onUpdate }) => {
   const handleLowPieceChange = (newValue: number) => {
     setMatch(
       setDetails(
-        alliance === 'red' ? 'redTeleLowPieces' : 'blueTeleLowPieces',
+        alliance === 'red' ? 'redAutoLowPieces' : 'blueAutoLowPieces',
         newValue
       )
     );
@@ -67,64 +109,71 @@ const TeleScoreSheet: FC<Props> = ({ alliance, participants, onUpdate }) => {
   const getChargeStatus = (station: number): number | undefined => {
     switch (station) {
       case 11:
-        return match.details?.redTeleChargeOne;
+        return match.details?.redAutoChargeOne;
       case 12:
-        return match.details?.redTeleChargeTwo;
+        return match.details?.redAutoChargeTwo;
       case 13:
-        return match.details?.redTeleChargeThree;
+        return match.details?.redAutoChargeThree;
       case 21:
-        return match.details?.blueTeleChargeOne;
+        return match.details?.blueAutoChargeOne;
       case 22:
-        return match.details?.blueTeleChargeTwo;
+        return match.details?.blueAutoChargeTwo;
       case 23:
-        return match.details?.blueTeleChargeThree;
+        return match.details?.blueAutoChargeThree;
       default:
         return 0;
     }
   };
 
-  const handleLinksChange = (newValue: number) => {
-    setMatch(
-      setDetails(alliance === 'red' ? 'redLinks' : 'blueLinks', newValue)
-    );
-  };
-
   const updateChargeStatus = (station: number, value: number) => {
     switch (station) {
       case 11:
-        setMatch(setDetails('redTeleChargeOne', value));
+        setMatch(setDetails('redAutoChargeOne', value));
         break;
       case 12:
-        setMatch(setDetails('redTeleChargeTwo', value));
+        setMatch(setDetails('redAutoChargeTwo', value));
         break;
       case 13:
-        setMatch(setDetails('redTeleChargeThree', value));
+        setMatch(setDetails('redAutoChargeThree', value));
         break;
       case 21:
-        setMatch(setDetails('blueTeleChargeOne', value));
+        setMatch(setDetails('blueAutoChargeOne', value));
         break;
       case 22:
-        setMatch(setDetails('blueTeleChargeTwo', value));
+        setMatch(setDetails('blueAutoChargeTwo', value));
         break;
       case 23:
-        setMatch(setDetails('blueTeleChargeThree', value));
+        setMatch(setDetails('blueAutoChargeThree', value));
         break;
     }
   };
 
-  const handleCoopertition = (newValue: number) => {
-    setMatch(setDetails('coopertitionBonus', newValue));
-  };
-
   return (
     <Grid container spacing={3}>
+      {participants?.map((p) => {
+        const update = (value: number) => {
+          updateMobility(p.station, value);
+        };
+
+        return (
+          <Grid item key={`${p.teamKey}-mobility`} xs={12} md={3} lg={3}>
+            <StateToggle
+              title={`${p.teamKey} Mobility`}
+              states={['No Mobility', 'Mobility']}
+              value={getMobility(p.station) ?? 0}
+              onChange={update}
+              fullWidth
+            />
+          </Grid>
+        );
+      })}
       <Grid item xs={12} md={4} lg={4}>
         <Typography variant='h6'>Top Game Pieces</Typography>
         <NumberInput
           value={
             alliance === 'red'
-              ? match.details.redTeleTopPieces
-              : match.details.blueTeleTopPieces
+              ? match.details.redAutoTopPieces
+              : match.details.blueAutoTopPieces
           }
           onChange={handleTopPieceChange}
         />
@@ -134,8 +183,8 @@ const TeleScoreSheet: FC<Props> = ({ alliance, participants, onUpdate }) => {
         <NumberInput
           value={
             alliance === 'red'
-              ? match.details.redTeleMidPieces
-              : match.details.blueTeleMidPieces
+              ? match.details.redAutoMidPieces
+              : match.details.blueAutoMidPieces
           }
           onChange={handleMidPieceChange}
         />
@@ -145,8 +194,8 @@ const TeleScoreSheet: FC<Props> = ({ alliance, participants, onUpdate }) => {
         <NumberInput
           value={
             alliance === 'red'
-              ? match.details.redTeleLowPieces
-              : match.details.blueTeleLowPieces
+              ? match.details.redAutoLowPieces
+              : match.details.blueAutoLowPieces
           }
           onChange={handleLowPieceChange}
         />
@@ -168,30 +217,8 @@ const TeleScoreSheet: FC<Props> = ({ alliance, participants, onUpdate }) => {
           </Grid>
         );
       })}
-      <Grid item xs={12} md={3} lg={3}>
-        <StateToggle
-          title={`Coopertition Bonus`}
-          states={['None', 'Cooperated']}
-          value={match.details.coopertitionBonus}
-          onChange={handleCoopertition}
-          fullWidth
-        />
-      </Grid>
-      <Grid item xs={12} md={4} lg={4}>
-        <Typography variant='h6'>
-          {alliance.toUpperCase()}&nbsp; Links
-        </Typography>
-        <NumberInput
-          value={
-            alliance === 'red'
-              ? match.details.redLinks
-              : match.details.blueLinks
-          }
-          onChange={handleLinksChange}
-        />
-      </Grid>
     </Grid>
   );
 };
 
-export default TeleScoreSheet;
+export default AutoScoreSheet;
