+++ front-end/src/seasons/ChargedUp/referee/ScoreSheet.tsx
@@ -5,13 +5,13 @@ import Typography from '@mui/material/Typography';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
 import { Alliance, ChargedUpDetails, Match } from '@toa-lib/models';
-import ConnectionChip from 'src/components/ConnectionChip/ConnectionChip';
-import MatchChip from 'src/components/MatchChip/MatchChip';
+import ConnectionChip from '@components/ConnectionChip/ConnectionChip';
+import MatchChip from '@components/MatchChip/MatchChip';
 import { useRecoilValue } from 'recoil';
-import { matchInProgressAtom } from 'src/stores/NewRecoil';
+import { matchInProgressAtom } from '@stores/NewRecoil';
 import TeamSheet from './TeamSheet';
 import PenaltySheet from './PenaltySheet';
-import TabPanel from 'src/components/TabPanel/TabPanel';
+import TabPanel from '@components/TabPanel/TabPanel';
 import AutoScoreSheet from './AutoScoreSheet';
 import TeleScoreSheet from './TeleOpScoreSheet';
 import { useSocket } from 'src/api/SocketProvider';
