+++ front-end/src/seasons/HydrogenHorizons/index.tsx
@@ -7,6 +7,7 @@ import {
 } from '..';
 import { Grid, Typography } from '@mui/material';
 import { isHydrogenHorizonsDetails } from '@toa-lib/models/build/seasons/HydrogenHorizons';
+import RefereeScoreSheet from '@seasons/HydrogenHorizons/referee/ScoreSheet';
 
 const MatchDetailInfo: FC<
   MatchDetailInfoProps<HydrogenHorizons.MatchDetails>
@@ -86,5 +87,6 @@ export const hydrogenHorizonComponents: SeasonComponents<HydrogenHorizons.MatchD
   {
     MatchDetailInfo,
     RedScoreBreakdown,
-    BlueScoreBreakdown
+    BlueScoreBreakdown,
+    RefereeScoreSheet
   };
